Phase B11: create tests/support/ with shared cli_e2e helpers
Extract all helper code from the 6788-line cli_e2e.rs into a five-file
support module: process.rs (server/mock/CLI runners), ports.rs, fs.rs
(manifest helpers + init state assertions), graphql.rs, and waits.rs
(polling helpers). Each sub-module is under 300 lines; support/mod.rs
re-exports all items for glob import.
